Move TDX bits
