+++ openhcl/openhcl_attestation_protocol/Cargo.toml
@@ -11,7 +11,7 @@ open_enum.workspace = true
 guid.workspace = true
 mesh.workspace = true
 sev_guest_device.workspace = true
-tdx_guest_device = { workspace = true, features = ["std"] }
+x86defs.workspace = true
 
 base64.workspace = true
 base64-serde.workspace = true
