+++ support/tdx_guest_device/Cargo.toml
@@ -6,16 +6,12 @@ name = "tdx_guest_device"
 edition.workspace = true
 rust-version.workspace = true
 
-[features]
-std = []
-
 [dependencies]
-bitfield-struct.workspace = true
-static_assertions.workspace = true
 zerocopy.workspace = true
 [target.'cfg(target_os = "linux")'.dependencies]
 nix = { workspace = true, features = ["ioctl"] }
 thiserror.workspace = true
+x86defs.workspace = true
 
 [lints]
 workspace = true
