+++ vm/x86/x86defs/Cargo.toml
@@ -11,6 +11,7 @@ open_enum.workspace = true
 
 arbitrary = { workspace = true, optional = true }
 bitfield-struct.workspace = true
+static_assertions.workspace = true
 zerocopy.workspace = true
 [lints]
 workspace = true
